some fixes + tweaks

+++ commons-asset-tooling/src/main/java/io/rocketbase/commons/tooling/ImageSourceHandling.java
@@ -7,17 +7,19 @@ import io.rocketbase.commons.service.DownloadService;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import net.coobird.thumbnailator.Thumbnails;
+import org.apache.commons.io.IOUtils;
 import org.apache.tika.Tika;
 
 import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.FileInputStream;
 
-public final class ImageHandling {
+public final class ImageSourceHandling {
 
     public static final DefaultDownloadService DEFAULT_DOWNLOAD_SERVICE = new DefaultDownloadService();
     private static final Tika TIKA = new Tika();
 
-    public static ImageHandlingBuilder builder(String url) {
+    public static ImageHandlingBuilder start(String url) {
         return new ImageHandlingBuilder(url);
     }
 
@@ -31,36 +33,40 @@ public final class ImageHandling {
             throw new RuntimeException("Source is not an image!");
         }
 
-        Thumbnails.Builder<File> thumbBuilder = Thumbnails.of(tempDownload.getFile());
-        if (width == null || height == null) {
-            if (width != null) {
-                thumbBuilder.width(width);
+        if (width != null || height != null) {
+            Thumbnails.Builder<File> thumbBuilder = Thumbnails.of(tempDownload.getFile());
+            if (width == null || height == null) {
+                if (width != null) {
+                    thumbBuilder.width(width);
+                } else {
+                    thumbBuilder.height(height);
+                }
             } else {
-                thumbBuilder.height(height);
+                thumbBuilder.scale(width, height);
             }
+            ByteArrayOutputStream thumbOs = new ByteArrayOutputStream();
+            thumbBuilder.toOutputStream(thumbOs);
+
+            // delete temp file
+            tempDownload.getFile().delete();
+
+            return new ImageHandlingResult(thumbOs.toByteArray(), assetType);
         } else {
-            thumbBuilder.scale(width, height);
+            return new ImageHandlingResult(IOUtils.toByteArray(new FileInputStream(tempDownload.getFile())), assetType);
         }
-        ByteArrayOutputStream thumbOs = new ByteArrayOutputStream();
-        thumbBuilder.toOutputStream(thumbOs);
-
-        // delete temp file
-        tempDownload.getFile().delete();
-
-        return new ImageHandlingResult(thumbOs, assetType);
     }
 
     @RequiredArgsConstructor
     public static class ImageHandlingResult {
-        private final ByteArrayOutputStream outputStream;
+        private final byte[] binary;
         private final AssetType assetType;
 
         public String base64() {
-            return "data:" + assetType.getContentType() + ";base64," + BaseEncoding.base64().encode(outputStream.toByteArray());
+            return "data:" + assetType.getContentType() + ";base64," + BaseEncoding.base64().encode(binary);
         }
 
-        public ByteArrayOutputStream stream() {
-            return outputStream;
+        public byte[] binary() {
+            return binary;
         }
 
         public AssetType assetType() {
@@ -70,7 +76,7 @@ public final class ImageHandling {
     }
 
     @RequiredArgsConstructor
-    private static class ImageHandlingBuilder {
+    public static class ImageHandlingBuilder {
         private final String url;
         private Integer width;
         private Integer height;
@@ -94,8 +100,8 @@ public final class ImageHandling {
             return this;
         }
 
-        public ImageHandlingResult build() {
-            return ImageHandling.build(downloadService, url, width, height);
+        public ImageHandlingResult process() {
+            return ImageSourceHandling.build(downloadService, url, width, height);
         }
     }
 }
